part_2/requirements.txt
pytest
requests
json-checker
pyyaml